requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.8.2
email-validator==2.2.0
python-dotenv==1.0.1
openai==1.51.2
httpx==0.27.2






